pdf_analyzer/requirements.txt
streamlit
pymupdf
openai
python-dotenv
streamlit-pdf-viewer
chromadb
langchain
langchain-openai
langchain-community
langchain-text-splitters
sentence-transformers
tiktoken